requirements.txt
absl-py==1.4.0
altair==4.2.2
astunparse==1.6.3
attrs==19.3.0
Automat==0.8.0
backports.zoneinfo==0.2.1
blinker==1.4
cachetools==5.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==23.1.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.7
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
flatbuffers==23.3.3
fonttools==4.39.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==6.3.0
importlib-resources==5.12.0
incremental==16.10.1
jax==0.4.8
Jinja2==3.1.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keras==2.12.0
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==16.0.0
macaroonbakery==1.3.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.0.1
mdurl==0.1.2
ml-dtypes==0.1.0
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.23.0
oauthlib==3.1.0
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==1.4.3
pexpect==4.6.0
Pillow==9.5.0
protobuf==3.20.3
pyarrow==11.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycups==1.9.73
pydeck==0.8.0
Pygments==2.15.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
Pympler==1.0.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rich==13.3.3
rsa==4.9
scipy==1.10.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
smmap==5.0.0
sos==4.4
ssh-import-id==5.10
streamlit==1.21.0
systemd-python==234
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
Twisted==18.9.0
typing-extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
ubuntu-advantage-tools==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
validators==0.20.0
wadllib==1.3.3
watchdog==3.0.0
Werkzeug==2.2.3
wrapt==1.14.1
xkit==0.0.0
zipp==3.15.0
zope.interface==4.7.1
